requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-telegram-bot==22.1
python-dotenv==1.1.1
pydantic==2.11.7
openai==1.99.9
scikit-learn==1.7.1
numpy==2.3.2
joblib==1.5.1
pytest==8.4.1
